Major refactor to swap drizzle for prisma and put api interactions in proper packages
