+++ apps/api-server/package.json
@@ -18,7 +18,6 @@
     "dist/**"
   ],
   "dependencies": {
-    "@prisma/client": "^6.10.1",
     "@repo/database": "workspace:*",
     "@trpc/server": "11.0.0-rc.477",
     "axios": "^1.10.0",
