+++ apps/api-server/sync-worker/index.ts
@@ -1,0 +1,2 @@
+import { syncSbirMetadataRecords } from "../src/lib/sbir";
+
