+++ apps/api-server/tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "extends":  "../../packages/tsconfig/base.json",
   "compilerOptions": {
     "strictNullChecks": true,
     "declaration": true,
@@ -6,7 +7,8 @@
     "rootDir": "./src",
     "baseUrl": "./src",
     "paths" : {
-      "@/*": ["./*"]
+      "@/*": ["./*"],
+      "@repo/database": ["../../packages/database/src/index"]
     },
     "resolveJsonModule": true,
     "strict": true,
