--- packages/database/drizzle/meta/_journal.json
@@ -1,13 +1,0 @@
-{
-  "version": "7",
-  "dialect": "postgresql",
-  "entries": [
-    {
-      "idx": 0,
-      "version": "7",
-      "when": 1744381277016,
-      "tag": "0000_cool_korg",
-      "breakpoints": true
-    }
-  ]
-}
