+++ packages/sbir-api/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@repo/sbir-api",
+  "version": "0.1.0",
+  "main": "src/index.ts",
+  "dependencies": {
+    "axios": "^1.10.0"
+  },
+  "devDependencies": {
+    "@types/node": "^20.19.1"
+  }
+}
